json_script: use size_t for calloc_a() length argument
json_script_file_from_blobmsg() passed name_len, a plain int, as a
chunk size to calloc_a(). __calloc_a() reads chunk sizes as size_t
through varargs, so on LP64 systems va_arg(ap, size_t) reads 8 bytes
where only 4 were pushed. Widen name_len to size_t.

Link: https://github.com/openwrt/libubox/pull/46
(cherry picked from commit 19e88cc41288478ae01d9246065ad319f9ca4e8e)

+++ json_script.c
@@ -39,7 +39,7 @@ json_script_file_from_blobmsg(const char *name, void *data, int len)
 {
 	struct json_script_file *f;
 	char *new_name;
-	int name_len = 0;
+	size_t name_len = 0;
 
 	if (name)
 		name_len = strlen(name) + 1;
